check pdf exists and remove comments

+++ bin/gather_pdf_notes.sh
@@ -19,15 +19,12 @@ copy_pdf_file() {
 	courseName=$(echo $(realpath --relative-to "$2" "$1") | \
 		awk -F/ '{print $1}')
 	pdfFile=$(echo "$1" | awk -F. '{print $1}')".pdf"
-	# echo $(dirname "$pdfFile")
-	# echo 
 	fileName=$(basename $(dirname "$pdfFile"))
-	# echo "$fileName"
-	# echo
 	destinationFile="$3/${courseName}_${fileName}.pdf"
-	echo "$pdfFile" "$destinationFile"	
-	cp "$pdfFile" "$destinationFile"
-	# echo "Copied $pdfFile to $destinationFile"
+	if [ -e "$pdfFile" ]; then
+		# tex file has been compiled 
+		cp "$pdfFile" "$destinationFile"
+	fi
 
 }
 
@@ -35,9 +32,5 @@ export -f copy_pdf_file
 
 find "$sourcePath" -type f -name '*.tex' | \
 	xargs -P 10 -I {} bash -c 'copy_pdf_file "$@"' _ {} $sourcePath $destPath
-	# while IFS= read -r sourceFile; do 
-	# copy_file $sourceFile	
-#done
 
-echo "All note files copied to $outPath"
